refactor(scripts): split version_bump_check.py by function cluster (P9-NEW)
Roadmap P9-NEW: split the 1,034-line version_bump_check.py CI gate
along its already-clear Verdict/Surface/heuristics/apply/render
boundaries into focused sibling modules, paired with a matching split
of the 1,150-line test_gates.py.

This is a pure mechanical extraction — function/class bodies are moved
verbatim. version_bump_check.py stays the byte-identical CLI entrypoint
at the same path: same args, same stdout, same exit codes, same
report/apply/hint semantics. It imports and re-exports every public
symbol so external importers (skill_sync_check.py, test_gates.py,
test_version_bump_check_extra.py) keep `from version_bump_check import`
working unchanged.

New modules (line counts):
  version_bump_surfaces.py    321  domain model, config loading,
                                   version-file I/O
  version_bump_heuristics.py  344  git-diff helpers, conv-commit
                                   classification, assess_surfaces
  version_bump_apply.py       114  bump arithmetic + apply_bumps
  version_bump_render.py      127  render_report
  gate_test_support.py        223  shared throwaway-repo Fixture
  test_version_bump_surfaces.py    131
  test_version_bump_heuristics.py  298
  test_version_bump_apply.py       145
  test_version_bump_fixfeat.py     343
  test_skill_sync.py               134

Before/after:
  version_bump_check.py  1034 -> 392 (CLI + fix/feat guard + main)
  test_gates.py          1150 ->  51 (aggregator importing all
                                      cluster TestCase subclasses)

The fix/feat-needs-bump guard stays in version_bump_check.py — it is
tightly coupled to the CLI and main(). test_gates.py remains the CI
entrypoint (.github/workflows/version-skill-check.yml runs
`python3 tools/scripts/test_gates.py`); it imports the five cluster
test modules so unittest discovers all 41 tests as before. Each
cluster test module is also independently runnable.

To keep test_version_bump_check_extra.py's `mock.patch.object(vbc, ...)`
stubs effective after the split, the cluster pipeline functions resolve
patchable git/whitespace/version helpers through the version_bump_check
entrypoint module at call time (deferred import, cycle-free).

Verification: entrypoint path and behavior byte-identical (--mode=report
dry-run stdout + exit code identical before/after via git stash; --help
byte-identical; classify-subject / hint / apply modes exercised); full
test_gates.py green (41 tests, same count as before); skill_sync_check.py
still runs (imports intact); test_version_bump_check_extra.py (41) and
test_gate_common.py (25) green.
